move utils/win_network.py tests to pytest
